refactor: 계층 단순화를 통한 props drilling 해소
